fix: 시험 관리 페이지 네비게이션 통일 및 최고점수 버그 수정
- 네비게이션 탭을 ExamsNavigationTabs로 분리하여 전 서브 페이지에 적용
- 각 서브 페이지 헤더에서 뒤로가기 버튼 제거
- 채점 결과 모달 최고점수 0 표시 버그 수정 (nullish coalescing → falsy 체크)
