+++ src/app/(dashboard)/educators/exams/page.tsx
@@ -3,6 +3,7 @@
 import { useSetBreadcrumb } from "@/hooks/useBreadcrumb";
 
 import { ExamsHeader } from "./_components/ExamsHeader";
+import { ExamsNavigationTabs } from "./_components/ExamsNavigationTabs";
 import { ExamsStats } from "./_components/ExamsStats";
 import { ExamsList } from "./_components/ExamsList";
 import { useExamsPage } from "./_hooks/useExamsPage";
@@ -15,6 +16,7 @@ export default function ExamsPage() {
   return (
     <div className="container mx-auto space-y-8 p-6">
       <ExamsHeader exams={exams} />
+      <ExamsNavigationTabs />
       <ExamsStats exams={exams} isLoading={isLoading} />
       <ExamsList
         exams={exams}
