+++ src/app/(dashboard)/educators/exams/report/page.tsx
@@ -2,6 +2,8 @@
 
 import { useSetBreadcrumb } from "@/hooks/useBreadcrumb";
 
+import { ExamsNavigationTabs } from "../_components/ExamsNavigationTabs";
+
 import { ReportHeader } from "./_components/ReportHeader";
 import { ReportSearchSection } from "./_components/ReportSearchSection";
 import { ReportTemplateSelect } from "./_components/ReportTemplateSelect";
@@ -17,6 +19,7 @@ export default function ReportPage() {
   return (
     <div className="container mx-auto space-y-6 p-6 xl:space-y-7">
       <ReportHeader />
+      <ExamsNavigationTabs />
 
       <div className="grid gap-6 xl:grid-cols-[360px_1fr]">
         <div className="space-y-6">
